Move page layout to it's own component
