+++ docs/js/page_layout.markdown
@@ -1,0 +1,8 @@
+
+
+<!-- Start src/components/page_layout.js -->
+
+## React
+
+<!-- End src/components/page_layout.js -->
+
